Migrate the docs site from Jekyll to Docusaurus (#208)
* Update .gitignore

Remove Jekyll-related folders and files. Add Docusaurus-related folders and files.

* Delete Jekyll-related file

* Move file

Move file for Docusaurus migration.

* Create versions.json

* Create favicon.ico

* Create scalar-logo.png

* Create .nojekyll

* Create custom.css

* Create index.js

* Create index.module.css

* Create styles.module.css

* Create index.js

* Create package.json

* Create babel.config.js

* Change invalid symbol

The arrow syntax used causes Docusaurus not to build, whereas a double-sided arrow symbol does not cause an issue.

* Delete file

Delete file since it's been moved to the i18n (internationalization) folder.

* Update versions.json

* Move file to proper folder

* Move file to proper folder

* Move file to proper folder

* Move file to proper folder

* Move file to proper folder

* Translate categories to Japanese

* Delete Jekyll-related file

* Create scalardl-logo.png

* Update image alt text

* Upgrade Docusaurus from `3.2.0` to `3.2.1`

* Delete file

Remove this file since we currently don't need it for our use case.

* Create sidebars.js

* Configure site settings and plugins

* Create release-notes.md

* Create release-notes.md

* Create release-notes.md

* Create release-notes.md

* Create release-notes.md

* Create unsupported-versions.md

* Create release-support-policy.md

* Create release-support-policy.md

* Create unsupported-versions.md

* Create 3.8 version

Since ScalarDL 3.9 has been released and is now the current version, we need to move the 3.8 docs to the `versioned_docs` folder.

* Remove unnecessary comment

* Add ScalarDL 3.9

* Remove 3.8 docs

Remove 3.8 docs since they've been generated in the `versioned_docs` folder.

* Add ScalarDL３.9 docs

* Move file

* Create doc

* Add 3.8

* Remove numbers used in testing

Remove the numbers used to test that the sidebar works as expected for versioned Japanese docs.

* Create version-3.8.json

* Change the current (latest) version to 3.9

* Change the default colors to Scalar blue

* Create _index.js

This file is currently not being used. If we want to use it in the future, remove the underline from the beginning of the file name and fix the error that occurs when trying to build the site.

* Create _index.js

This file is currently not being used. If we want to use it in the future, remove the underline from the beginning of the file name and fix the error that occurs when trying to build the site.

* Revise wording

* Add release notes for version

* Move file

* Fix heading 1 formatting

* Add `@docusaurus/plugin-content-docs`

* Add updated Helm Chart doc

* Remove `latest` from version title

Not exactly necessary and adds more work for when a new version is released.

* Add missing translation

* Add updated Kubernetes doc

* Add new Kubernetes doc

* Add new Kubernetes docs; capitalize categories

* Create version-3.8-sidebars.json

* Create version-3.7-sidebars.json

* Create version-3.6-sidebars.json

* Create version-3.5-sidebars.json

* Create version-3.4-sidebars.json

* Remove unnecessary numbering

* Delete unnecessary doc

This doc existed only because we previously didn't have collapsible sections in the sidebar nav. Now that we have collapsible sections, this doc isn't needed.

* Change tab syntax from HTML to MDX-supported format

* Change file type from `.md` to `.mdx`

This change is necessary to support tabs (and other features that render as React components).

* Remove out-of-support banner

The banner is applied to all docs for this version in the docusaurus.config.js file.

* Add support for syncing tab choices

* Delete `docusaurus-lunr-search` plugin

* Delete unnecessary, commented out configurations

* Move misplaced docs

* Add configurations for Algolia DocSearch

* Change locale format for Japanese docs

* Add redirect for docs at current `ja-jp` path

On the current Jekyll-based docs site, the path to Japanese docs is `/docs/ja-jp`. However, in Docusaurus, the path is (naturally) `/ja-jp/docs`.

* Enable "Edit this page" button

* Enable breadcrumbs at the top of pages

* Modify location of item in footer

* Add doc

* Add license key configuration doc

* Reduce page title (h1) font size

* Add `Docs` link and fix links in footer

* Add updated doc

* Add updated doc

* Add new doc

* Add links to new Kubernetes and Helm Charts docs

* Add locale to language

* Remove unnecessary comment

* Translate category titles

* Mark versions as unsupported

* Remove unneeded dropdown item

* Fix redirects for base URL of Japanese docs

* Delete file

This file has been moved but was added back in when merging commits from the `main` branch.

* Fix note syntax

Change the note syntax that was mistakenly added back in when merging commits from the `main` branch.

* Change note placement

Move note below title so that the title renders properly in the sidebar nav. If the note is at the top of the file, the title in the sidebar nav becomes the file name (in this case, "AzureMarketplaceGuide").

* Create _translation-ja-jp.mdx

This is an include that tells readers of Japanese docs that the contents of the page has been machine translated and to refer to the English version of the doc if there are any discrepancies.

* Change file extension from `.md` to `.mdx`

* Change `.md` links to `.mdx`

* Remove old code block syntax

Remove old code block syntax since it's not needed in Docusaurus.

* Add translation banner to top of doc