Add Release Support Policy page redirect; fix odd broken link behavior on Release Notes pages (#245)
* Add redirect for release support policy page

* Change links to absolute links

For some reason, using relative links breaks links in this doc. When a visitor clicks a link to a release note, they are taken to the correct release note page. However, if they click the back button to go back to this page and then click a link to a release note, the link is broken because the URL to the release note is incorrect. Changing the relative links to absolute links solves the issue.