Add quotation marks to versions (#253)
Docusaurus throws an error and cannot build the site when a minor version reaches `X.XX`. In our case, the error displayed is `unknown version (3.1) found`. To get around this when we release an `X.XX` version in the future, we need to add double quotation marks around version numbers.