Add support for tags in docs header (#685)
* Add Font Awesome plugins

* Add styles for Font Awesome question-mark icon

* Add styles for tags at top of page

Add styles to put tags on the same line as `Version`. Also add styles to make the tags look decent when the dark theme is applied on the docs site.

* Create `index.tsx` (swizzled)

* Add tags to top of page

Make tags more likely to be seen at first glance on a page by adding components here.

* Comment out `Editions` tag header; uncomment tag syntax

* Create index.tsx (swizzle)

* Comment out tags in footer

Since we want to have tags at the top of the page we should hide them from the footer.

* Render both Enterprise edition tags as `Enterprise` (#686)

Make `Enterprise Standard` and `Enterprise Premium` tags appear as `Enterprise`.