Add support dropdown menu to right-hand sidebar of pages (#733)
* Swizzle component to add support dropdown menu

This component has been swizzled so that we can add a support dropdown menu.

* Add `@typebot.io/react`

* Create AssistantModal.tsx

This is called when the user clicks the AI Assistant button in the Support dropdown menu.

* Add Support dropdown menu above TOC

* Add Support button styles

* Create SupportDropdownMenu.tsx

This contains the dropdown items.

* Change dropdown item font color

Change dropdown item font color to match the default color of other fonts in Docusaurus.

* Hide support dropdown menu from home page