Add links to overview pages for first-level and second-level categories in sidebar navigation (#860)
* Add style to shrink category arrows in sidebar nav

Add a style to make the arrow in the sidebar navigation smaller. This is necessary because making only some categories clickable indexes causes the icon to revert to the default size.

* Make level 1 and 2 category overview pages into category landing pages

By making category overview pages the landing pages for those categories, we can direct visitors directly to those sections and reduce the number of items shown in the categories by one.

* Move backup and restore docs