Redesign version home pages (#914)
* Create scalardl-presentation-overview.png

* Create home-page-background.avif

* Add styles for version home page redesign

* Add `react-lite-youtube-embed` dependency

* Redesign version home page component

* Update version home page component

* Change `new content` wording to `recent features`

Since the docs shown on the `new content` line won't always be new for that version of docs that the visitor is reading, `recent features` is more appropriate wording here.

* Add docs to "Recent features" section

* Change link for diversity's sake

* Change `new content` wording to `recent features`

Since the docs shown on the `new content` line won't always be new for that version of docs that the visitor is reading, `recent features` is more appropriate wording here.

* Revise comment about links in the `Recent features` section

* Delete index.tsx

* Add comment about link guidelines

* Disable zooming on images with `youtube-embed` style

We currently have an image that zooms in when it is clicked on. Since we'll eventually have a video on the home page, I've added a style here to not zoom on the image on the home page (or any other image with the style `youtube-embed`).

* Remove awkward, extra padding below home page image

Since we're currently using an image since we don't have a YouTube video to add to the home page, we need to remove the extra padding caused by using an image

* Add ScalarDL presentation overview screenshot and link

Eventually, we'll add a YouTube video about ScalarDL here. Until we have one, we've decided to use this presentation for visitors to learn more.